Ability to take a photo and add a photo via the selector (WIP)
Can take a photo via the camera and recieve the image data. This is then saved on the filesystem and a simple record in the db to show relation to entity.

Can add photo(s) via the photo selector. Similar as above, these are stored on the filesystem and then saved to db.

Implementation still required additional abstractions and refactorings.