Further improvements and refactorings of the Photo Gallery
Move to using a recycler view, which should help with performance of large datasets.
Also moved the buttons into the action bar, a design that will be used throughout.

Attempted some performance improvements when rendering bitmaps by sampling them down to the required ImageView size, however I cannot seem to get the calculated height and width in time (always 0). So for now it is set to a constant 250x250, which should be better than the original photo size.