Refactor handling of bitmaps and their rendering
Use AsyncTask to achieve decoding of bitmaps and updating of the UI elements once done. This alleviates a number of slow downs with multiple bitmaps on screen.

Collapsing bar in holiday activity now uses ImageView instead of background which enables the ability to crop and have the same image as in the list view.

Style holidays list using CardView better. This is the standard to be used in rest of app.